# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "h12 API - V1",
    "description": "“Like for post” entity was added\n\nAn authorized user can like the post",
    "version": "h12"
  },
  "paths": {
...
    "/hometask_12/api/blogs": {
      "get": {
        "tags": ["Blogs"],
        "summary": "Returns blogs with paging",
        "parameters": [
          {
            "name": "searchNameTerm",
            "in": "query",
            "description": "Search term for blog Name: Name should contains this term in any position",
            "schema": {
              "type": "string",
              "default": null
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Default value: desc",
            "schema": {
              "$ref": "#/components/schemas/h04.SortDirections"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "pageNumber is number of portions that should be returned",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize is portions size that should be returned",
            "schema": {
              "maximum": 20,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h03.BlogViewModel>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h03.BlogViewModel>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h03.BlogViewModel>"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Blogs"],
        "summary": "Create new blog",
        "requestBody": {
          "description": "Data for constructing new Blog entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogInputModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created blog",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h03.BlogViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h03.BlogViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h03.BlogViewModel"
                }
              }
            }
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      }
    },
    "/hometask_12/api/blogs/{blogId}/posts": {
      "get": {
        "tags": ["Blogs"],
        "summary": "Returns all posts for specified blog",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "pageNumber is number of portions that should be returned",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize is portions size that should be returned",
            "schema": {
              "maximum": 20,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Default value: desc",
            "schema": {
              "$ref": "#/components/schemas/h04.SortDirections"
            }
          },
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              }
            }
          },
          "404": {
            "description": "If specificied blog is not exists"
          }
        }
      },
      "post": {
        "tags": ["Blogs"],
        "summary": "Create new post for specific blog",
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Data for constructing new post entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogPostInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogPostInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogPostInputModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created post",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              }
            }
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "If specified blog doesn't exists"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      }
    },
    "/hometask_12/api/blogs/{id}": {
      "get": {
        "tags": ["Blogs"],
        "summary": "Returns blog by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Existing blog id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h03.BlogViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h03.BlogViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h03.BlogViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": ["Blogs"],
        "summary": "Update existing Blog by id with InputModel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Data for updating",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogInputModel"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      },
      "delete": {
        "tags": ["Blogs"],
        "summary": "Delete blog specified by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Blog id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIErrorResult": {
        "type": "object",
        "properties": {
          "errorsMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message with error explanation for certain field",
            "nullable": true
          },
          "field": {
            "type": "string",
            "description": "What field/property of input model has error",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Paginator<h03.BlogViewModel>": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "pagesCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/h03.BlogViewModel"
            }
          }
        },
        "additionalProperties": false
      },
...
      "Paginator<h12.PostViewModel>": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "pagesCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/h12.PostViewModel"
            }
          }
        },
        "additionalProperties": false
      },
      "h02.db.BlogInputModel": {
        "required": ["description", "name", "websiteUrl"],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 15,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "type": "string"
          },
          "websiteUrl": {
            "maxLength": 100,
            "pattern": "^https://([a-zA-Z0-9_-]+\\.)+[a-zA-Z0-9_-]+(\\/[a-zA-Z0-9_-]+)*\\/?$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "h02.db.BlogPostInputModel": {
        "required": ["content", "shortDescription", "title"],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 30,
            "type": "string"
          },
          "shortDescription": {
            "maxLength": 100,
            "type": "string"
          },
          "content": {
            "maxLength": 1000,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "h03.BlogViewModel": {
        "required": ["description", "id", "name", "websiteUrl"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "websiteUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isMembership": {
            "type": "boolean",
            "description": "True if user has not expired membership subscription to blog"
          }
        },
        "additionalProperties": false
      },
      "h04.SortDirections": {
        "enum": ["asc", "desc"],
        "type": "string"
      },
...
      "h11.LikeStatus": {
        "enum": ["None", "Like", "Dislike"],
        "type": "string",
        "description": "Send None if you want to unlike\\undislike"
      },
...
      "h12.ExtendedLikesInfoViewModel": {
        "type": "object",
        "properties": {
          "likesCount": {
            "type": "integer",
            "description": "Total likes for parent item",
            "format": "int32"
          },
          "dislikesCount": {
            "type": "integer",
            "description": "Total dislikes for parent item",
            "format": "int32"
          },
          "myStatus": {
            "$ref": "#/components/schemas/h11.LikeStatus"
          },
          "newestLikes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/h12.LikeDetailsViewModel"
            },
            "description": "Last 3 likes (status \"Like\")",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "h12.LikeDetailsViewModel": {
        "type": "object",
        "properties": {
          "addedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "login": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Details about single like"
      },
      "h12.PostViewModel": {
        "required": [
          "blogId",
          "blogName",
          "content",
          "id",
          "shortDescription",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "blogId": {
            "type": "string"
          },
          "blogName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "extendedLikesInfo": {
            "$ref": "#/components/schemas/h12.ExtendedLikesInfoViewModel"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Enter JWT Bearer token **_only_**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "refreshToken": {
        "type": "apiKey",
        "description": "JWT refreshToken inside cookie. Must be correct, and must not expire",
        "name": "refreshToken",
        "in": "cookie"
      }
    }
  }
}